Refactor hibernate objects
